+++ fake-cucumber/javascript/src/cli.ts
@@ -2,11 +2,12 @@ import { Command } from 'commander'
 import packageJson from '../package.json'
 import loadSupportCode from './loadSupportCode'
 import runCucumber from './runCucumber'
+import os from 'os'
 import { GherkinStreams, IGherkinStreamOptions } from '@cucumber/gherkin-streams'
 import { Query as GherkinQuery } from '@cucumber/gherkin-utils'
 import { version } from '../package.json'
 import * as messages from '@cucumber/messages'
-import createMeta from '@cucumber/create-meta'
+import detectCiEnvironment from '@cucumber/ci-environment'
 import { MessageToNdjsonStream } from '@cucumber/message-streams'
 import { RunOptions } from './types'
 
@@ -40,7 +41,25 @@ async function main() {
   envelopeOutputStream.pipe(process.stdout)
 
   const metaEnvelope: messages.Envelope = {
-    meta: createMeta('fake-cucumber', version, process.env),
+    meta: {
+      protocolVersion: messages.version,
+      implementation: {
+        version,
+        name: 'fake-cucumber',
+      },
+      cpu: {
+        name: os.arch(),
+      },
+      os: {
+        name: os.platform(),
+        version: os.release(),
+      },
+      runtime: {
+        name: 'node.js',
+        version: process.versions.node,
+      },
+      ci: detectCiEnvironment(process.env),
+    },
   }
   envelopeOutputStream.write(metaEnvelope)