+++ gherkin/go/cmd/main.go
@@ -9,8 +9,8 @@ import (
 	"encoding/json"
 	"flag"
 	"fmt"
-	"github.com/cucumber/common/gherkin/go/v22"
-	"github.com/cucumber/common/messages/go/v17"
+	"github.com/cucumber/common/gherkin/go/v23"
+	"github.com/cucumber/common/messages/go/v18"
 	"os"
 )