+++ json-formatter/go/message_lookup.go
@@ -3,7 +3,7 @@ package json
 import (
 	"fmt"
 
-	"github.com/cucumber/common/messages/go/v17"
+	"github.com/cucumber/common/messages/go/v18"
 )
 
 type MessageLookup struct {