+++ react/javascript/src/GherkinQueryContext.ts
@@ -1,4 +1,0 @@
-import React from 'react'
-import { Query as GherkinQuery } from '@cucumber/gherkin-utils'
-
-export default React.createContext(new GherkinQuery())