+++ react/javascript/src/components/gherkin/statusName.ts
@@ -1,5 +1,0 @@
-import * as messages from '@cucumber/messages'
-
-export default (status: messages.TestStepResultStatus): string => {
-  return status.toLowerCase()
-}