+++ react/javascript/src/components/index.ts
@@ -1,3 +1,0 @@
-export * as app from './app'
-export * as gherkin from './gherkin'
-export * as customise from './customise'